Fixes when a enum value is not in the provided list or is added to the response from Emma. Will fall back to Unknown until a value is created on this end. Also updates all id fields to long (int64) since the application seems to be extending past the int32 limit on some fields.
This closes #42 and closes #40.

+++ EmmaSharp/Models/Groups/Group.cs
@@ -29,14 +29,14 @@ namespace EmmaSharp.Models.Groups
         public GroupType GroupType { get; set; }
 
         [JsonProperty("member_group_id")]
-        public int? MemberGroupId { get; set; }
+        public long? MemberGroupId { get; set; }
 
         [JsonConverter(typeof(EmmaDateConverter))]
         [JsonProperty("purged_at")]
         public DateTime? PurgedAt { get; set; }
 
         [JsonProperty("account_id")]
-        public int? AccountId { get; set; }
+        public long? AccountId { get; set; }
 
         [JsonProperty("group_name")]
         public string GroupName { get; set; }
