+++ EmmaSharp/Models/Members/Member.cs
@@ -16,7 +16,7 @@ namespace EmmaSharp.Models.Members
         public DateTime? ConfirmedOptIn { get; set; }
 
         [JsonProperty("account_id")]
-        public int? AccountId { get; set; }
+        public long? AccountId { get; set; }
 
         [JsonProperty("fields")]
         public Dictionary<string, object> Fields { get; set; }
